requirements.txt
# Core
discord.py>=2.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.0

# Data / Math
numpy>=1.26.0
pytz>=2024.1
yfinance>=0.2.30
scipy>=1.12.0

# Charts
matplotlib>=3.8.0

# Optional: faster event loop (Linux only, recommended for AWS)
uvloop>=0.19.0; sys_platform != "win32"
